requirements.txt
# GUI Framework
PySide6>=6.7,<7

# AI Integration
google-generativeai>=0.8.0
anthropic>=0.40.0
ollama>=0.2.0
httpx>=0.27.0

# Tooling & Analysis
radon>=6.0.0
rope>=1.12.0
GitPython>=3.1.0
SQLAlchemy>=2.0.0
packaging>=23.0

# Testing
pytest>=8.0.0
rich
